Move spark-elastic to backend module
